Move protocol pkg at server root
